added title and about page
