+++ public/about.html
@@ -29,7 +29,7 @@
   </head>
 
   <body>
-    <a class="title" href="/about" target="_blank">EDITOR</a>
+    <a class="title" href="/about">EDITOR</a>
     <main></main>
     <script src="/index.js"></script>
   </body>
